--- app/src/ui/LensList/TopToolbar/TopToolbar.tsx
@@ -1,17 +1,0 @@
-import { Box, Typography } from '@mui/material'
-import Distance from './Distance'
-import UnitsToggle from './UnitsToggle'
-
-export default function TopToolbar() {
-    return (
-        <Box display="flex">
-            <Box flexGrow={1}>
-                <Typography variant="h2">Lenses</Typography>
-            </Box>
-            <Box mr={1}>
-                <Distance />
-            </Box>
-            <UnitsToggle />
-        </Box>
-    )
-}
