+++ app/src/ui/LensTable/Table/Row/Aperture.tsx
