+++ app/src/ui/LensTable/Table/Row/Name.tsx
@@ -1,9 +1,11 @@
 import { TextField } from '@mui/material'
 import { ChangeEvent } from 'react'
 import useDoFStore from '../../../../store'
+import useIsMobile from '../../../../utilities/useIsMobile'
 
 export default function Name({ lens }: { lens: LensDefinition }) {
     const { updateLens } = useDoFStore()
+    const isMobile = useIsMobile()
     const onChange = (event: ChangeEvent<HTMLInputElement>) => {
         updateLens({ ...lens, name: event.target.value })
     }
@@ -16,6 +18,7 @@ export default function Name({ lens }: { lens: LensDefinition }) {
             autoComplete="off"
             size="small"
             data-testid={`lens-name-${lens.id}`}
+            fullWidth={isMobile}
         />
     )
 }
