+++ app/src/ui/LensTable/Table/Row/Sensor.tsx
